extrats two logo components
